fix: fix mypy errors for SortedDict's types
The main change is to have SortedDict subclass MutableMapping rather than
dict. This is necessary because dict's type differs from MutableMapping
subtly in that the type it returns from keys() has an additional
property that is not present on KeysView that MutableMapping returns
from keys(). It seems to be recommended to subclass MutableMapping
instead of dict to avoid this.
More here: https://bugs.python.org/issue46399

Also a few insignificant changes to make the methods compatible with
MutableMapping.