fix: broaden typing of __new__ for pyright compatibility
pyright (used by VSCode's pylance) does not allow/support our __new__
signatures that only use a TypeVar in the return position, such as
() -> SortedList[_OrderT] for SortedList. That's meant to constrain the
created list to only contain orderable elements, and this works with
mypy. However it seems that using a TypeVar only in the return position
is not a standardised feature of typing specs, so support varies between
implementations. To be pragmatic, these __new__ signatures which tried
to constrain to orderable types are now typed to allow Any (implicitly,
by returning Self).