test: add test for SortedKeyList(key=None) bug
SortedKeyList constructor is incorrectly typed to allow `key` to be None.
In fact `key` must be a callable, None is not converted to the identity
function, despite the default key value being identity if not provided.

This adds a test for `SortedKeyList(key=None)` being invalid.

See: https://github.com/h4l/sortedcontainers-stubs/issues/12