fix: make `SortedList(key_fn)` a type error
SortedList() constructor was incorrectly accepting a positional `key`
arg when it should have been a type error. It's now typed to
only allow a `key` with no `iterable` if the `key` is a keyword
argument.

Incidentally, this change has fixed an unrelated mypy issue that caused
subscripting the constructor to explicitly declare the type to not work.
`SortedList[xxx]()` now works as expected.

See: https://github.com/h4l/sortedcontainers-stubs/issues/10