fix: re-type `SortedKeyList()` to match implementation
SortedKeyList() constructor was typed to accept its `key` argument as a
positional argument in the first position, which was incorrect, as the
`iterable` argument is always the first positional argument. It also
allowed `key` to be None, but the implementation requires it to always be a callable.

See: https://github.com/h4l/sortedcontainers-stubs/issues/10
See: https://github.com/h4l/sortedcontainers-stubs/issues/12